--- a/pkg/tagset/hash_generator.go
+++ b/pkg/tagset/hash_generator.go
@@ -140,7 +140,7 @@ func (g *HashGenerator) Hash(tb *HashingTagsAccumulator) uint64 {
 // tag each tag is present once in either l or r, but not both at the same time.
 //
 // First, duplicates are removed from l. Then duplicates are removed from r, including any tags that
-// are already present in l. Can move tags from one accumulator to another.
+// are already present in l.
 func (g *HashGenerator) Dedup2(l *HashingTagsAccumulator, r *HashingTagsAccumulator) {
 	ntags := l.Len() + r.Len()
 
@@ -154,9 +154,9 @@ func (g *HashGenerator) Dedup2(l *HashingTagsAccumulator, r *HashingTagsAccumula
 	//    less than 16 tags
 	//  - n > hashSetSize: sort
 	if ntags > hashSetSize {
-		l.AppendHashingAccumulator(r)
 		l.SortUniq()
-		r.Reset()
+		r.SortUniq()
+		r.removeSorted(l)
 	} else if ntags > bruteforceSize {
 		// reset the `seen` hashset.
 		// it copies `g.empty` instead of using make because it's faster