--- a/pkg/tagset/hashing_tags_accumulator_test.go
+++ b/pkg/tagset/hashing_tags_accumulator_test.go
@@ -69,3 +69,33 @@ func TestHashingTagsAccumulatorCopy(t *testing.T) {
 	assert.Equal(t, []string{"test", "b", "c"}, tagsCopy)
 	assert.Equal(t, []string{"a", "b", "c"}, tb.data)
 }
+
+func TestRemoveSorted(t *testing.T) {
+	l := NewHashingTagsAccumulator()
+	r := NewHashingTagsAccumulator()
+	l.Append("a", "b", "c", "d")
+	r.Append("a", "b", "e", "f")
+	r.removeSorted(l)
+	assert.Equal(t, []string{"a", "b", "c", "d"}, l.Get())
+	assert.ElementsMatch(t, []string{"e", "f"}, r.Get())
+
+	r.Reset()
+	r.Append("c", "d", "e", "f")
+	r.removeSorted(l)
+	assert.ElementsMatch(t, []string{"e", "f"}, r.Get())
+
+	r.Reset()
+	r.Append("a", "aa", "ab", "b")
+	r.removeSorted(l)
+	assert.ElementsMatch(t, []string{"aa", "ab"}, r.Get())
+
+	r.Reset()
+	r.Append("A", "B", "a", "d")
+	r.removeSorted(l)
+	assert.ElementsMatch(t, []string{"A", "B"}, r.Get())
+
+	r.Reset()
+	r.Append("A", "a", "b", "e")
+	r.removeSorted(l)
+	assert.ElementsMatch(t, []string{"A", "e"}, r.Get())
+}