### What does this PR do?

When there were more than 512 tags (prior to deduplication) Dedup2
would move client tags into the tagger buffer so they can deduplicated.

We rely on tagger tags now for context telemetry, and it is confusing
if that telemetry contains client tags. It is bad for cardinality too.

Instead of combining two slices and doing SortUniq on them, we now
SortUniq each buffer individually, and add another
pass (`removeSorted`) to remove tagger tags from the client tags, if
any present.

### Motivation

Avoid client tags showing up in dogstatsd_contexts_by_origin metric if the client sends too many repeated tags.

### Additional Notes

<!--
* Anything else we should know when reviewing?
* Include benchmarking information here whenever possible.
* Include info about alternatives that were considered and why the proposed
  version was chosen.
-->

### Possible Drawbacks / Trade-offs

```
benchmark                        old ns/op     new ns/op     delta
BenchmarkDedup2/1-tags-36        10.6          10.9          +2.82%
BenchmarkDedup2/2-tags-36        13.0          13.0          +0.00%
BenchmarkDedup2/4-tags-36        20.2          20.4          +0.79%
BenchmarkDedup2/8-tags-36        30.9          30.9          +0.00%
BenchmarkDedup2/16-tags-36       45.0          48.0          +6.76%
BenchmarkDedup2/32-tags-36       70.8          77.2          +8.98%
BenchmarkDedup2/64-tags-36       128           146           +14.26%
BenchmarkDedup2/128-tags-36      247           293           +18.82%
BenchmarkDedup2/256-tags-36      494           570           +15.42%
BenchmarkDedup2/512-tags-36      9501          9730          +2.41%
BenchmarkDedup2/1024-tags-36     9760          9793          +0.34%
BenchmarkDedup2/2048-tags-36     19116         22742         +18.97%
```

### Describe how to test/QA your changes

<!--
* Write here in detail or link to detailed instructions on how this change can
  be tested/QAd/validated, including any environment setup.
-->

### Reviewer's Checklist
<!--
* Authors can use this list as a reference to ensure that there are no problems
  during the review but the signing off is to be done by the reviewer(s).

Note: Adding GitHub labels is only possible for contributors with write access.
-->

- [ ] If known, an appropriate milestone has been selected; otherwise the `Triage` milestone is set.
- [ ] Use the `major_change` label if your change either has a major impact on the code base, is impacting multiple teams or is changing important well-established internals of the Agent. This label will be use during QA to make sure each team pay extra attention to the changed behavior. For any customer facing change use a releasenote.
- [ ] A [release note](https://github.com/DataDog/datadog-agent/blob/main/docs/dev/contributing.md#reno) has been added or the `changelog/no-changelog` label has been applied.
- [ ] Changed code has automated tests for its functionality.
- [ ] Adequate QA/testing plan information is provided if the `qa/skip-qa` label is not applied.
- [ ] At least one `team/..` label has been applied, indicating the team(s) that should QA this change.
- [ ] If applicable, docs team has been notified or [an issue has been opened on the documentation repo](https://github.com/DataDog/documentation/issues/new).
- [ ] If applicable, the `need-change/operator` and `need-change/helm` labels have been applied.
- [ ] If applicable, the `k8s/<min-version>` label, indicating the lowest Kubernetes version compatible with this feature. 
- [ ] If applicable, the [config template](https://github.com/DataDog/datadog-agent/blob/main/pkg/config/config_template.yaml) has been updated.